Add giraph init

+++ giraph/src/in/ds256/Giraph/OutDegreeComputation.java
@@ -4,6 +4,7 @@ import org.apache.giraph.Algorithm;
 import org.apache.giraph.graph.BasicComputation;
 import org.apache.giraph.graph.Vertex;
 import org.apache.hadoop.io.DoubleWritable;
+import org.apache.hadoop.io.FloatWritable;
 import org.apache.hadoop.io.LongWritable;
 
 import java.io.IOException;
@@ -15,11 +16,11 @@ import java.io.IOException;
         name = "Outdegree Count"
 )
 public class OutDegreeComputation extends BasicComputation<
-        LongWritable, LongWritable, DoubleWritable, DoubleWritable> {
+        LongWritable, DoubleWritable, FloatWritable, DoubleWritable> {
 
     @Override
-    public void compute(Vertex<LongWritable, LongWritable, DoubleWritable> vertex, Iterable<DoubleWritable> messages) throws IOException {
-        LongWritable vertexValue = vertex.getValue();
+    public void compute(Vertex<LongWritable, DoubleWritable, FloatWritable> vertex, Iterable<DoubleWritable> messages) throws IOException {
+        DoubleWritable vertexValue = vertex.getValue();
         vertexValue.set(vertex.getNumEdges());
         vertex.setValue(vertexValue);
         vertex.voteToHalt();
